Refatorado os testes de produto

+++ src/main/java/models/request/produto/ProdutoRequest.java
@@ -3,13 +3,15 @@ package models.request.produto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ProdutoRequest {
 
     public String nome;
     public String descricao;
-    public Double preco;
+    public BigDecimal preco;
     public Integer estoque;
     public String categoria;
 }
